feat: add the possibility to specify references
This can be useful if you write notes about a book, podcast, blog
article(s), etc and you want to generally references them. That doesn't
mean you can't cite them in the markdown file, it just means it's easier
to use those references for task like search because they're in the
front matter.